chore: pin jsonschema version to 4.17.3
Fixes issue where versions 4.18 and later depend on rpds-py which requires Rust compilation in the buildsystem. Tests on python versions 3.11 and 3.12 were not passing, likely because there weren't any pre-built binaries for these versions in the environment. Alternatively we can add support for Rust in the CI to build rpds-py if we want to keep the latest version of jsonschema.

See for more details: https://github.com/python-jsonschema/jsonschema/issues/1117

FMEHUBDEV-1610